add test client, rename createL1Client, and move to opstack test utils
