+++ test/src/opStack.ts
@@ -1,7 +1,10 @@
-import { type Chain, optimism } from '~viem/chains/index.js'
+import { type Chain, optimism, mainnet } from '~viem/chains/index.js'
 import { createClient } from '~viem/clients/createClient.js'
+import { createPublicClient } from '~viem/clients/createPublicClient.js'
 import { http } from '~viem/clients/transports/http.js'
 import { accounts, warn } from './constants.js'
+import { createHttpServer } from './utils.js'
+import { createTestClient } from '~viem/index.js'
 
 export let anvilPortOptimism: number
 if (process.env.VITE_ANVIL_PORT_OPTIMISM) {
@@ -63,3 +66,40 @@ export const optimismClientWithAccount = createClient({
 export const optimismClientWithoutChain = createClient({
   transport: http(localHttpUrlOptimism),
 })
+
+/**
+ * @experimental
+ * A simple test utility that simply starts and stops a tevm server running mainnet
+ * l1 devnet with Optimism mainnet contracts deployed. Also returns a public client
+ * Can be nuked in favor of using anvil once OP stack contracts are deployed
+ */
+export const createL1Server = async () => {
+  const { createHttpHandler } = await import('@tevm/server')
+  const { createL1Client } = await import('@tevm/opstack')
+
+  const evmClient = createL1Client()
+  await evmClient.ready()
+
+  const handler = createHttpHandler(evmClient)
+
+  const server = await createHttpServer(handler)
+
+  const publicClient = createPublicClient({
+    chain: mainnet,
+    transport: http(server.url),
+  })
+
+  const testClient = createTestClient({
+    mode: 'anvil',
+    transport: http(server.url),
+    chain: mainnet,
+  })
+
+  // return only the server
+  // returning the tevm client would be useful so we could interact with it in tests
+  // but we want these tests to not be coupled to tevm. Thus they should only
+  // interact with the server via http JSON-RPC using eth_ and anvil_ methods
+  return { server, publicClient, testClient }
+}
+
+/* c8 ignore stop */
